+++ packages/create-docusaurus/src/utils.ts
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+/**
+ * We use a simple kebab-case-like conversion
+ * It's not perfect, but good enough
+ * We don't want to depend on lodash in this package
+ * See https://github.com/facebook/docusaurus/pull/11653
+ * @param siteName
+ */
+export function siteNameToPackageName(siteName: string): string {
+  const match = siteName.match(
+    /[A-Z]{2,}(?=[A-Z][a-z]+\d*|\b|_)|[A-Z]?[a-z]+\d*|[A-Z]|\d+/g,
+  );
+  if (match) {
+    return match.map((x) => x.toLowerCase()).join('-');
+  }
+  return siteName;
+}
