+++ packages/docusaurus-utils/src/index.ts
@@ -37,7 +37,7 @@ export {
   getPluginI18nPath,
   getLocaleConfig,
 } from './i18nUtils';
-export {mapAsyncSequential, findAsyncSequential, kebabCase} from './jsUtils';
+export {mapAsyncSequential, findAsyncSequential} from './jsUtils';
 export {
   normalizeUrl,
   getEditUrl,
