+++ packages/docusaurus-utils/src/jsUtils.ts
@@ -43,14 +43,3 @@ export async function findAsyncSequential<T>(
   }
   return undefined;
 }
-
-/**
- * Converts a string to kebab-case.
- * Mimics lodash's behavior.
- */
-export function kebabCase(str: string): string {
-  const match = str.match(
-    /[A-Z]{2,}(?=[A-Z][a-z]+[0-9]*|\b|[_])|[A-Z]?[a-z]+[0-9]*|[A-Z]|[0-9]+/g,
-  );
-  return match ? match.map((x) => x.toLowerCase()).join('-') : '';
-}
